pyproject.toml
[project]
name = "fine-tuning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "accelerate>=1.13.0",
    "peft>=0.19.1",
    "pillow>=12.2.0",
    "qwen-vl-utils>=0.0.14",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "transformers>=5.6.2",
]
